# openapi/pvlog-v1.yaml
openapi: 3.1.1
...
info:
  title: PVLog REST API
  version: 0.1.0
  summary: Modern self-hosted photovoltaic monitoring API
  description: >-
    The versioned JSON contract for PVLog. PVOutput r2 compatibility endpoints
    are documented separately because they use legacy form and CSV formats.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
servers:
  - url: /
    description: Current PVLog installation
...
paths:
...
  /api/v1/systems/{system_id}/observations:
    parameters: [{ $ref: "#/components/parameters/NestedSystemId" }]
    post:
      operationId: createObservation
      summary: Ingest one observation
      description: Normalizes and durably stores one idempotent canonical observation.
      tags: [Telemetry]
      security: [{ bearerAuth: [telemetry:write] }]
      parameters: [{ name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1 } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ObservationInput" }, example: { observedAtEpochMillis: 1780000000000, generationPowerWatts: 4200, generationEnergyWh: 350 } } } }
      responses: { "201": { description: Observation accepted, content: { application/json: { schema: { type: object } } } }, "422": { $ref: "#/components/responses/ValidationProblem" }, "503": { $ref: "#/components/responses/OverloadProblem" } }
  /api/v1/systems/{system_id}/observations/batch:
    parameters: [{ $ref: "#/components/parameters/NestedSystemId" }]
    post:
      operationId: createObservationBatch
      summary: Ingest an observation batch
      description: Processes a bounded batch in atomic or partial-result mode.
      tags: [Telemetry]
      security: [{ bearerAuth: [telemetry:write] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ObservationBatchInput" } } } }
      responses: { "200": { description: Stable indexed batch outcomes, content: { application/json: { schema: { type: object } } } }, "422": { $ref: "#/components/responses/ValidationProblem" }, "503": { $ref: "#/components/responses/OverloadProblem" } }
  /api/v1/systems/{system_id}/observations/{observation_id}:
    parameters: [{ $ref: "#/components/parameters/NestedSystemId" }, { $ref: "#/components/parameters/ObservationId" }]
    patch:
      operationId: correctObservation
      summary: Correct an observation
      description: Applies an optimistic replacement or deletion overlay and schedules reconciliation.
      tags: [Telemetry]
      security: [{ bearerAuth: [telemetry:write] }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [expectedVersion, reason, delete], properties: { expectedVersion: { type: integer, minimum: 1 }, reason: { type: string, minLength: 1 }, generationPowerWatts: { type: integer }, delete: { type: boolean } } } } } }
      responses: { "200": { description: Immediately visible merged observation, content: { application/json: { schema: { type: object } } } }, "409": { $ref: "#/components/responses/ConflictProblem" } }
    delete:
      operationId: deleteObservation
      summary: Delete an observation
      description: Records an optimistic deletion while preserving audit and reconciliation history.
      tags: [Telemetry]
      security: [{ bearerAuth: [telemetry:write] }]
      parameters: [{ $ref: "#/components/parameters/IfMatch" }, { name: X-Correction-Reason, in: header, required: true, schema: { type: string, minLength: 1 } }]
      responses: { "204": { description: Deletion accepted }, "409": { $ref: "#/components/responses/ConflictProblem" } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    NestedSystemId: { name: system_id, in: path, required: true, schema: { $ref: "#/components/schemas/UuidV7" } }
    ObservationId: { name: observation_id, in: path, required: true, schema: { $ref: "#/components/schemas/UuidV7" } }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string }, example: '"3"' }
...
  responses:
...
    ValidationProblem: { description: Request validation failed, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
...
    ConflictProblem: { description: Optimistic concurrency conflict, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
    OverloadProblem: { description: Ingestion is temporarily saturated, headers: { Retry-After: { schema: { type: integer, minimum: 1 } } }, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
...
  schemas:
...
    ObservationInput:
      type: object
      required: [observedAtEpochMillis]
      properties: { observedAtEpochMillis: { type: integer }, generationPowerWatts: { type: integer }, generationEnergyWh: { type: integer }, consumptionPowerWatts: { type: integer }, consumptionEnergyWh: { type: integer }, voltageMillivolts: { type: integer, minimum: 0 }, temperatureMillidegreesCelsius: { type: integer }, sourceReference: { type: string } }
    ObservationBatchInput:
      type: object
      required: [mode, items]
      properties: { mode: { enum: [atomic, partial] }, items: { type: array, minItems: 1, maxItems: 1000, items: { allOf: [{ $ref: "#/components/schemas/ObservationInput" }, { type: object, required: [idempotencyKey], properties: { idempotencyKey: { type: string, minLength: 1 } } }] } } }
    UuidV7:
      type: string
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
      description: >-
        Canonical lowercase UUID version 7 generated by PVLog. Treat this value
        as opaque; its time-sortable bits do not grant authority or establish an
        authoritative creation timestamp.
      examples:
        - 019505c8-7c85-7f0b-9bc3-2a3c4d5e6f70
...
    Problem:
      type: object
      additionalProperties: true
      description: RFC 9457 problem details.
      required:
        - type
        - title
        - status
        - requestId
      properties:
        type:
          type: string
          format: uri-reference
          description: Stable problem type identifier.
        title:
          type: string
          description: Short human-readable problem title.
        status:
          type: integer
          minimum: 400
          maximum: 599
          description: HTTP status code.
        detail:
          type: string
          description: Safe occurrence-specific explanation.
        requestId:
          $ref: "#/components/schemas/UuidV7"